rename "constants" to "enums"
