+++ src/components/FoodItem.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect } from 'react'
 import { type Entry } from '../types/Entry'
-import { FoodStatus } from '../constants/FoodStatus'
+import { FoodStatus } from '../enums/FoodStatuses'
 import EditFoodButton from './EditFoodButton'
 import DeleteFoodButton from './DeleteFoodButton'
