--- src/constants/FoodLocation.ts
@@ -1,5 +1,0 @@
-export enum FoodLocation {
-  fridge = 'Fridge',
-  freezer = 'Freezer',
-  pantry = 'Pantry'
-}
