+++ src/enums/FoodLocations.ts
@@ -1,0 +1,5 @@
+export enum FoodLocations {
+  freezer = 'Freezer',
+  fridge = 'Fridge',
+  pantry = 'Pantry'
+}
